fix: resolve linter warnings and broken type inferences for iOS test mocks
- Removed obsolete @override annotation on iOS mock method
- Removed unnecessary local variable type annotations in better_player_android and better_player_ios
- Explicitly cast FFI dynamic map lookups (textureId, pos) to resolve analyzer errors

+++ packages/better_player_android/lib/better_player_android.dart
@@ -170,7 +170,7 @@ class BetterPlayerAndroid extends BetterPlayerPlatform {
     final player = createJniPlayer(callback);
     if (player == null) return null;
 
-    final textureId = player.textureId;
+    final textureId = player.textureId as int;
     _players[textureId] = createWrapper(player);
     _callbacks[textureId] = callback;
     _eventControllers[textureId] = StreamController<VideoEvent>.broadcast();
